Refactor search functionality in KnowledgeTable and SearchBox
